fix(shop): lift state up from ProductList
